Add files via upload
In this scenario, we have the same number of players for the two teams, namely two players in each team. Our players are black and the opponent players are red. The dynamic for the ball we used is as follows: the opponent player with the ball has a blue color and when our player (defensive player) gets close enough to the ball carrier player, he passes the ball to its teammate materialized by the change of color of its teammate from red to blue. Our goal is to prevent them from getting close to the goal or from scoring.